requirements.txt
matplotlib
pandas
psycopg2-binary
scipy
numpy
psycopg
pandas
numpy
scipy
python-dotenv